+++ tests/testthat/test-choices_selected.R
@@ -92,3 +92,50 @@ testthat::test_that("choices_selected throws when no_select_keyword is passed to
 testthat::test_that("is returns choices_selected if passed a choices selected object", {
   testthat::expect_equal(is(choices_selected(choices = "A")), "choices_selected")
 })
+
+testthat::test_that("choices_selected remove duplicates", {
+  testthat::expect_identical(
+    choices_selected(choices = c("A", "A")),
+    structure(list(choices = "A", selected = "A", fixed = FALSE), class = "choices_selected")
+  )
+  testthat::expect_identical(
+    choices_selected(choices = setNames(c("A", "A"), c("A", "A"))),
+    structure(list(choices = c(A = "A"), selected = c(A = "A"), fixed = FALSE),
+      class = "choices_selected"
+    )
+  )
+  testthat::expect_equal(
+    choices_selected(
+      choices = c(
+        "name for A" = "A", "name for A" = "A",
+        "Name for nothing" = "", "name for b" = "B", "name for C" = "C"
+      ),
+      selected = c("A", "A")
+    ),
+    structure(list(
+      choices = c(`name for A` = "A", `Name for nothing` = "", `name for b` = "B", `name for C` = "C"),
+      selected = "A", fixed = FALSE
+    ),
+    class = "choices_selected"
+    )
+  )
+
+  testthat::expect_equal(
+    choices_selected(
+      structure(c(`STUDYID: Study Identifier` = "STUDYID", `STUDYID: Study Identifier` = "STUDYID"),
+        raw_labels = c(STUDYID = "Study Identifier", STUDYID = "Study Identifier"),
+        combined_labels = c("STUDYID: Study Identifier", "STUDYID: Study Identifier"),
+        class = c("choices_labeled", "character")
+      )
+    ),
+    structure(list(
+      choices = structure(c(`STUDYID: Study Identifier` = "STUDYID"),
+        raw_labels = c(STUDYID = "Study Identifier"),
+        combined_labels = "STUDYID: Study Identifier",
+        class = c("choices_labeled", "character")
+      ),
+      selected = c(`STUDYID: Study Identifier` = "STUDYID"),
+      fixed = FALSE
+    ), class = "choices_selected")
+  )
+})